+++ crates/wire/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Remove useless quantification on `Yoke::{try_,}map()` function argument
 - Optimize memory usage when encoding slices in particular of small elements like `Vec<u8>`
 
 ### Patch