+++ crates/wire/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Patch
 
+- Factorize recursive schema logic
 - Update dependencies
 
 ## 0.1.2
@@ -38,4 +39,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 4 -->
+<!-- Increment to skip CHANGELOG.md test: 5 -->